requirements.txt
psycopg[binary]>=3.1.0
python-dotenv==1.0.0
google-generativeai>=0.3.0
neo4j>=5.14.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
langchain>=0.1.0
langchain-google-genai>=0.0.6
langgraph>=0.0.20
numpy>=1.24.0

